# backend/docs/swagger.yaml
basePath: /api
host: lifegate-backend.onrender.com
info:
  contact:
    email: support@lifegate.health
    name: LifeGate Support
  description: AI-powered clinical diagnostic platform. Patients interact with an
    AI assistant, physicians review AI-generated reports, and administrators manage
    the full lifecycle.
  license:
    name: Proprietary
  termsOfService: https://lifegate.health/terms
  title: LifeGate API
  version: "1.0"
paths:
  /admin/audit:
    get:
      parameters:
      - description: Filter by event type
        in: query
        name: eventType
        type: string
      - description: Filter by actor role
        in: query
        name: actorRole
        type: string
      - description: Filter by resource type
        in: query
        name: resource
        type: string
      - description: Start date (YYYY-MM-DD)
        in: query
        name: dateFrom
        type: string
      - description: End date (YYYY-MM-DD)
        in: query
        name: dateTo
        type: string
      - description: Page (default 1)
        in: query
        name: page
        type: integer
      - description: Items per page (default 50)
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                type: array
              meta:
                type: object
              success:
                type: boolean
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              message:
                type: string
              success:
                type: boolean
            type: object
      security:
      - BearerAuth: []
      summary: Audit log
      tags:
      - admin
  /admin/audit/export:
    get:
      parameters:
      - description: Filter by event type
        in: query
        name: eventType
        type: string
      - description: Filter by actor role
        in: query
        name: actorRole
        type: string
      - description: Filter by resource type
        in: query
        name: resource
        type: string
      - description: Start date (YYYY-MM-DD)
        in: query
        name: dateFrom
        type: string
      - description: End date (YYYY-MM-DD)
        in: query
        name: dateTo
        type: string
      produces:
      - text/csv
      responses:
        "200":
          description: CSV file download
          schema:
            type: file
        "500":
          description: Internal Server Error
          schema:
            properties:
              message:
                type: string
              success:
                type: boolean
            type: object
      security:
      - BearerAuth: []
      summary: Export audit log CSV
      tags:
      - admin
  /admin/cases:
    get:
      parameters:
      - description: Page (default 1)
        in: query
        name: page
        type: integer
      - description: Items per page (default 20)
        in: query
        name: pageSize
        type: integer
      - description: Filter by status
        in: query
        name: status
        type: string
      - description: Filter by urgency
        in: query
        name: urgency
        type: string
      - description: Filter by category
        in: query
        name: category
        type: string
      - description: Search by patient name or condition
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                type: array
              meta:
                properties:
                  page:
                    type: integer
                  pageSize:
                    type: integer
                  total:
                    type: integer
                type: object
              success:
                type: boolean
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              message:
                type: string
              success:
                type: boolean
            type: object
      security:
      - BearerAuth: []
      summary: Admin case list
      tags:
      - admin
  /admin/compliance/ndpa:
    get:
      parameters:
      - description: Max results (default 30)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                type: array
              success:
                type: boolean
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              message:
                type: string
              success:
                type: boolean
            type: object
      security:
      - BearerAuth: []
      summary: NDPA compliance snapshots
      tags:
      - admin
  /admin/compliance/ndpa/export:
    get:
      parameters:
      - description: Max records (default 90)
        in: query
        name: limit
        type: integer
      produces:
      - text/csv
      responses:
        "200":
          description: CSV file download
          schema:
            type: file
        "500":
          description: Internal Server Error
          schema:
            properties:
              message:
                type: string
              success:
                type: boolean
            type: object
      security:
      - BearerAuth: []
      summary: Export NDPA compliance CSV
      tags:
      - admin
  /admin/compliance/ndpa/generate:
    post:
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            properties:
              data:
                type: object
              success:
                type: boolean
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              message:
                type: string
              success:
                type: boolean
            type: object
      security:
      - BearerAuth: []
      summary: Generate NDPA snapshot
      tags:
      - admin
  /admin/dashboard:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                type: object
              success:
                type: boolean
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              message:
                type: string
              success:
                type: boolean
            type: object
      security:
      - BearerAuth: []
      summary: Admin dashboard stats
      tags:
      - admin
...
  /admin/physicians:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                type: array
              success:
                type: boolean
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              message:
                type: string
              success:
                type: boolean
            type: object
      security:
      - BearerAuth: []
      summary: List all physicians
      tags:
      - admin
    post:
      consumes:
      - application/json
      parameters:
      - description: Physician data
        in: body
        name: body
        required: true
        schema:
          properties:
            email:
              type: string
            name:
              type: string
            password:
              type: string
            phone:
              type: string
            specialization:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            properties:
              data:
                properties:
                  id:
                    type: string
                type: object
              message:
                type: string
              success:
                type: boolean
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              message:
                type: string
              success:
                type: boolean
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              message:
                type: string
              success:
                type: boolean
            type: object
      security:
      - BearerAuth: []
      summary: Create physician account
      tags:
      - admin
  /admin/physicians/{id}:
    delete:
      parameters:
      - description: Physician user ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              message:
                type: string
              success:
                type: boolean
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              message:
                type: string
              success:
                type: boolean
            type: object
      security:
      - BearerAuth: []
      summary: Delete physician
      tags:
      - admin
    get:
      parameters:
      - description: Physician user ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                type: object
              success:
                type: boolean
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              message:
                type: string
              success:
                type: boolean
            type: object
      security:
      - BearerAuth: []
      summary: Get physician detail
      tags:
      - admin
    patch:
      consumes:
      - application/json
      parameters:
      - description: Physician user ID
        in: path
        name: id
        required: true
        type: string
      - description: Fields to update
        in: body
        name: body
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              message:
                type: string
              success:
                type: boolean
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              message:
                type: string
              success:
                type: boolean
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              message:
                type: string
              success:
                type: boolean
            type: object
      security:
      - BearerAuth: []
      summary: Update physician
      tags:
      - admin
  /admin/physicians/{id}/mdcn-override:
    post:
      consumes:
      - application/json
      parameters:
      - description: Physician user ID
        in: path
        name: id
        required: true
        type: string
      - description: 'Status: confirmed | rejected'
        in: body
        name: body
        required: true
        schema:
          properties:
            status:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              message:
                type: string
              success:
                type: boolean
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              message:
                type: string
              success:
                type: boolean
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              message:
                type: string
              success:
                type: boolean
            type: object
      security:
      - BearerAuth: []
      summary: Override MDCN verification
      tags:
      - admin
  /admin/physicians/{id}/suspend:
    post:
      consumes:
      - application/json
      parameters:
      - description: Physician user ID
        in: path
        name: id
        required: true
        type: string
      - description: Suspension reason
        in: body
        name: body
        schema:
          properties:
            reason:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              message:
                type: string
              success:
                type: boolean
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              message:
                type: string
              success:
                type: boolean
            type: object
      security:
      - BearerAuth: []
      summary: Suspend physician
      tags:
      - admin
  /admin/physicians/{id}/unsuspend:
    post:
      parameters:
      - description: Physician user ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              message:
                type: string
              success:
                type: boolean
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              message:
                type: string
              success:
                type: boolean
            type: object
      security:
      - BearerAuth: []
      summary: Unsuspend physician
      tags:
      - admin
  /admin/physicians/flag-check:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                properties:
                  newlyFlagged:
                    type: integer
                type: object
              success:
                type: boolean
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              message:
                type: string
              success:
                type: boolean
            type: object
      security:
      - BearerAuth: []
      summary: Trigger SLA flag check
      tags:
      - admin
  /admin/settings/alerts:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                type: array
              success:
                type: boolean
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              message:
                type: string
              success:
                type: boolean
            type: object
      security:
      - BearerAuth: []
      summary: Get alert thresholds
      tags:
      - admin
...
  /admin/sla:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                type: array
              success:
                type: boolean
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              message:
                type: string
              success:
                type: boolean
            type: object
      security:
      - BearerAuth: []
      summary: SLA report
      tags:
      - admin
...
  /genai/status:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                type: object
              success:
                type: boolean
            type: object
      security:
      - BearerAuth: []
      summary: AI provider status
      tags:
      - genai
...
schemes:
- https
securityDefinitions:
  BearerAuth:
    description: 'Enter your JWT token as: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"